+++ core/staking/stake_verifier_snapshot.go
@@ -173,7 +173,7 @@ func (s *StakeVerifier) OnStateLoaded(ctx context.Context) error {
 	}
 	if lastBlockSeen != 0 {
 		s.log.Info("restoring staking bridge starting block", logging.Uint64("block", lastBlockSeen))
-		s.ethEventSource.UpdateStakingStartingBlock(lastBlockSeen)
+		// s.ethEventSource.UpdateStakingStartingBlock(lastBlockSeen)
 	}
 	return nil
 }