+++ core/types/snapshot.go
@@ -103,6 +103,7 @@ const (
 	VolumeDiscountProgramSnapshot  SnapshotNamespace = "volumeDiscountProgram"
 	LiquidationSnapshot            SnapshotNamespace = "liquidation"
 	TxCacheSnapshot                SnapshotNamespace = "txCache"
+	EVMHeartbeatSnapshot           SnapshotNamespace = "evmheartbeat"
 
 	MaxChunkSize   = 16 * 1000 * 1000 // technically 16 * 1024 * 1024, but you know
 	IdealChunkSize = 10 * 1000 * 1000 // aim for 10MB